+++ core/src/main/kotlin/cl/ravenhill/keen/engine/OptimizationEngine.kt
@@ -5,12 +5,13 @@
 
 package cl.ravenhill.keen.engine
 
+import cl.ravenhill.keen.objective.Objective
 import cl.ravenhill.keen.repr.Feature
 
 
 interface OptimizationEngine<F>  where F : Feature<*, F> {
 
-    val objectiveFunction: (F) -> Double
+    val objective: Objective<F>
 
     fun optimize(initialState: F): F
 }